Finished exercises 4.15 to 4.23.

+++ controllers/blogs.js
@@ -1,27 +1,45 @@
 const blogsRouter = require("express").Router();
 const Blog = require("../models/blog");
+const { userExtractor } = require("../utils/middleware");
 
 blogsRouter.get("/", async (request, response) => {
-  const blogs = await Blog.find({});
+  const blogs = await Blog.find({}).populate("user", {
+    username: 1,
+    name: 1,
+    id: 1,
+  });
   response.json(blogs);
 });
 
-blogsRouter.post("/", async (request, response) => {
+blogsRouter.post("/", userExtractor, async (request, response) => {
   const blog = new Blog(request.body);
+
+  const user = request.user;
+
   if (blog.title === undefined || blog.url === undefined) {
     response.status(400).end();
   } else {
     if (blog.likes === undefined) {
       blog.likes = 0;
     }
+    blog.user = user.id;
     const savedBlog = await blog.save();
+    user.blogs = user.blogs.concat(savedBlog.id);
+    await user.save();
     response.status(201).json(savedBlog);
   }
 });
 
-blogsRouter.delete("/:id", async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id);
-  response.status(204).end();
+blogsRouter.delete("/:id", userExtractor, async (request, response) => {
+  const user = request.user;
+  const blog = await Blog.findById(request.params.id);
+
+  if (blog.user.toString() === user.id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id);
+    response.status(204).end();
+  } else {
+    return response.status(401).json({ error: "unauthorized deletion" });
+  }
 });
 
 blogsRouter.put("/:id", async (request, response) => {
@@ -31,10 +49,12 @@ blogsRouter.put("/:id", async (request, response) => {
     title: body.title,
     author: body.author,
     url: body.url,
-    likes: body.likes
+    likes: body.likes,
   };
 
-  const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, { new: true });
+  const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, {
+    new: true,
+  });
   response.json(updatedBlog);
 });
